requirements.txt
alembic==1.9.2
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==3.4.8
dnspython==2.2.1
email-validator==1.3.1
fastapi==0.89.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Mako==1.2.4
MarkupSafe==2.1.2
motor==3.1.1
passlib==1.7.4
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
sniffio==1.3.0
SQLAlchemy==2.0.0
SQLAlchemy-Utils==0.39.0
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
zipp==3.12.0
